01-metadata/Modbus/requirements.txt
RPi.GPIO
adafruit-circuitpython-bme680
pyModbusTCP